fix: refactor json-rpc
