+++ ethereum/rpc/api/addrlock.go
@@ -1,4 +1,4 @@
-package ethapi
+package api
 
 import (
 	"sync"
