+++ ethereum/rpc/api/dbapi.go
@@ -1,4 +1,4 @@
-package ethapi
+package api
 
 import (
 	"errors"
