+++ ethereum/rpc/apis.go
@@ -7,37 +7,33 @@ import (
 	"github.com/ethereum/go-ethereum/rpc"
 
 	"github.com/artela-network/artela/ethereum/rpc/api"
-	"github.com/artela-network/artela/ethereum/rpc/ethapi"
 	"github.com/artela-network/artela/ethereum/rpc/filters"
 )
 
 func GetAPIs(clientCtx client.Context, wsClient *rpcclient.WSClient, logger log.Logger, apiBackend *BackendImpl) []rpc.API {
-	nonceLock := new(ethapi.AddrLocker)
+	nonceLock := new(api.AddrLocker)
 	return []rpc.API{
 		{
 			Namespace: "eth",
-			Service:   ethapi.NewEthereumAPI(apiBackend, logger),
+			Service:   api.NewEthereumAPI(apiBackend, logger),
 		}, {
 			Namespace: "eth",
-			Service:   ethapi.NewBlockChainAPI(apiBackend, logger),
+			Service:   api.NewBlockChainAPI(apiBackend, logger),
 		}, {
 			Namespace: "eth",
-			Service:   ethapi.NewTransactionAPI(apiBackend, logger, nonceLock),
+			Service:   api.NewTransactionAPI(apiBackend, logger, nonceLock),
 		}, {
 			Namespace: "txpool",
-			Service:   ethapi.NewTxPoolAPI(apiBackend, logger),
-		}, {
-			Namespace: "debug",
-			Service:   ethapi.NewDebugAPI(apiBackend),
+			Service:   api.NewTxPoolAPI(apiBackend, logger),
 		}, {
 			Namespace: "debug",
 			Service:   api.NewDebugAPI(apiBackend),
 		}, {
 			Namespace: "eth",
-			Service:   ethapi.NewEthereumAccountAPI(apiBackend),
+			Service:   api.NewEthereumAccountAPI(apiBackend),
 		}, {
 			Namespace: "personal",
-			Service:   ethapi.NewPersonalAccountAPI(apiBackend, logger, nonceLock),
+			Service:   api.NewPersonalAccountAPI(apiBackend, logger, nonceLock),
 		}, {
 			Namespace: "net",
 			Service:   api.NewNetAPI(apiBackend),
