+++ ethereum/rpc/backend.go
@@ -4,20 +4,15 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"math"
 	"math/big"
-	"runtime"
 	"strconv"
 	"time"
 
 	sdkmath "cosmossdk.io/math"
-	bftclient "github.com/cometbft/cometbft/rpc/client"
 	tmrpctypes "github.com/cometbft/cometbft/rpc/core/types"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/server"
 	sdktypes "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/version"
-	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
@@ -31,27 +26,28 @@ import (
 	"github.com/ethereum/go-ethereum/params"
 	"github.com/ethereum/go-ethereum/rpc"
 
-	"github.com/artela-network/artela/ethereum/rpc/api"
-	"github.com/artela-network/artela/ethereum/rpc/ethapi"
-	ethapi2 "github.com/artela-network/artela/ethereum/rpc/ethapi"
+	"github.com/artela-network/artela/ethereum/rpc/backend"
 	"github.com/artela-network/artela/ethereum/rpc/filters"
 	rpctypes "github.com/artela-network/artela/ethereum/rpc/types"
-	"github.com/artela-network/artela/ethereum/rpc/utils"
 	"github.com/artela-network/artela/ethereum/server/config"
-	"github.com/artela-network/artela/ethereum/types"
 	ethereumtypes "github.com/artela-network/artela/ethereum/types"
 	"github.com/artela-network/artela/x/evm/txs"
-	"github.com/artela-network/artela/x/evm/txs/support"
-	evmtypes "github.com/artela-network/artela/x/evm/types"
 	feetypes "github.com/artela-network/artela/x/fee/types"
 )
 
 var (
 	_ gasprice.OracleBackend = (*BackendImpl)(nil)
-	_ ethapi2.Backend        = (*BackendImpl)(nil)
 	_ filters.Backend        = (*BackendImpl)(nil)
-	_ api.NetBackend         = (*BackendImpl)(nil)
-	_ api.DebugBackend       = (*BackendImpl)(nil)
+
+	_ backend.Backend             = (*BackendImpl)(nil)
+	_ backend.EthereumBackend     = (*BackendImpl)(nil)
+	_ backend.BlockChainBackend   = (*BackendImpl)(nil)
+	_ backend.TrancsactionBackend = (*BackendImpl)(nil)
+	_ backend.DebugBackend        = (*BackendImpl)(nil)
+	_ backend.PersonalBackend     = (*BackendImpl)(nil)
+	_ backend.TxPoolBackend       = (*BackendImpl)(nil)
+	_ backend.NetBackend          = (*BackendImpl)(nil)
+	_ backend.Web3Backend         = (*BackendImpl)(nil)
 )
 
 // backend represents the backend for the JSON-RPC service.
@@ -76,6 +72,107 @@ type BackendImpl struct {
 	ctx         context.Context
 	clientCtx   client.Context
 	queryClient *rpctypes.QueryClient
+}
+
+func (b *BackendImpl) CurrentHeader() (*ethtypes.Header, error) {
+	block, err := b.BlockByNumber(context.Background(), rpc.LatestBlockNumber)
+	if err != nil {
+		return nil, err
+	}
+	if block == nil || block.Header() == nil {
+		return nil, errors.New("current block header not found")
+	}
+	return block.Header(), nil
+}
+
+func (b *BackendImpl) Accounts() []common.Address {
+	addresses := make([]common.Address, 0) // return [] instead of nil if empty
+
+	infos, err := b.clientCtx.Keyring.List()
+	if err != nil {
+		b.logger.Info("keying list failed", "error", err)
+		return nil
+	}
+
+	for _, info := range infos {
+		pubKey, err := info.GetPubKey()
+		if err != nil {
+			b.logger.Info("getPubKey failed", "info", info, "error", err)
+			return nil
+		}
+		addressBytes := pubKey.Address().Bytes()
+		addresses = append(addresses, common.BytesToAddress(addressBytes))
+	}
+
+	return addresses
+}
+
+func (b *BackendImpl) GetBalance(address common.Address, blockNrOrHash rpc.BlockNumberOrHash) (*hexutil.Big, error) {
+	blockNum, err := b.blockNumberFromCosmos(blockNrOrHash)
+	if err != nil {
+		return nil, err
+	}
+
+	req := &txs.QueryBalanceRequest{
+		Address: address.String(),
+	}
+
+	_, err = b.CosmosBlockByNumber(blockNum)
+	if err != nil {
+		return nil, err
+	}
+
+	res, err := b.queryClient.Balance(rpctypes.ContextWithHeight(blockNum.Int64()), req)
+	if err != nil {
+		return nil, err
+	}
+
+	val, ok := sdkmath.NewIntFromString(res.Balance)
+	if !ok {
+		return nil, errors.New("invalid balance")
+	}
+
+	if val.IsNegative() {
+		return nil, errors.New("couldn't fetch balance. Node state is pruned")
+	}
+
+	return (*hexutil.Big)(val.BigInt()), nil
+}
+
+func (b *BackendImpl) ArtBlockByNumber(_ context.Context, number rpc.BlockNumber) (*rpctypes.Block, error) {
+	resBlock, err := b.CosmosBlockByNumber(number)
+	if err != nil || resBlock == nil {
+		return nil, fmt.Errorf("query block failed, block number %d, %w", number, err)
+	}
+
+	blockRes, err := b.CosmosBlockResultByNumber(&resBlock.Block.Height)
+	if err != nil {
+		return nil, fmt.Errorf("block result not found for height %d", resBlock.Block.Height)
+	}
+
+	return b.BlockFromCosmosBlock(resBlock, blockRes)
+}
+
+func (b *BackendImpl) BlockByHash(_ context.Context, hash common.Hash) (*rpctypes.Block, error) {
+	resBlock, err := b.CosmosBlockByHash(hash)
+	if err != nil || resBlock == nil {
+		return nil, fmt.Errorf("failed to get block by hash %s, %w", hash.Hex(), err)
+	}
+
+	blockRes, err := b.CosmosBlockResultByNumber(&resBlock.Block.Height)
+	if err != nil {
+		return nil, fmt.Errorf("block result not found for height %d", resBlock.Block.Height)
+	}
+
+	return b.BlockFromCosmosBlock(resBlock, blockRes)
+}
+
+func (b *BackendImpl) ChainConfig() *params.ChainConfig {
+	cfg, err := b.chainConfig()
+	if err != nil {
+		return nil
+	}
+	return cfg
 }
 
 func (b *BackendImpl) EthBlockByNumber(blockNum rpc.BlockNumber) (*ethtypes.Block, error) {
@@ -131,42 +228,6 @@ func (b *BackendImpl) SyncProgress() ethereum.SyncProgress {
 	}
 }
 
-func (b *BackendImpl) SuggestGasTipCap(baseFee *big.Int) (*big.Int, error) {
-	if baseFee == nil {
-		// london hardfork not enabled or feemarket not enabled
-		return big.NewInt(0), nil
-	}
-
-	params, err := b.queryClient.FeeMarket.Params(b.ctx, &feetypes.QueryParamsRequest{})
-	if err != nil {
-		return nil, err
-	}
-	// calculate the maximum base fee delta in current block, assuming all block gas limit is consumed
-	// ```
-	// GasTarget = GasLimit / ElasticityMultiplier
-	// Delta = BaseFee * (GasUsed - GasTarget) / GasTarget / Denominator
-	// ```
-	// The delta is at maximum when `GasUsed` is equal to `GasLimit`, which is:
-	// ```
-	// MaxDelta = BaseFee * (GasLimit - GasLimit / ElasticityMultiplier) / (GasLimit / ElasticityMultiplier) / Denominator
-	//          = BaseFee * (ElasticityMultiplier - 1) / Denominator
-	// ```t
-	maxDelta := baseFee.Int64() * (int64(params.Params.ElasticityMultiplier) - 1) / int64(params.Params.BaseFeeChangeDenominator) // #nosec G701
-	if maxDelta < 0 {
-		// impossible if the parameter validation passed.
-		maxDelta = 0
-	}
-	return big.NewInt(maxDelta), nil
-}
-
-func (b *BackendImpl) ChainConfig() *params.ChainConfig {
-	cfg, err := b.chainConfig()
-	if err != nil {
-		return nil
-	}
-	return cfg
-}
-
 func (b *BackendImpl) chainConfig() (*params.ChainConfig, error) {
 	params, err := b.queryClient.Params(b.ctx, &txs.QueryParamsRequest{})
 	if err != nil {
@@ -180,95 +241,6 @@ func (b *BackendImpl) chainConfig() (*params.ChainConfig, error) {
 	}
 
 	return params.Params.ChainConfig.EthereumConfig(currentHeader.Number.Int64(), b.chainID), nil
-}
-
-func (b *BackendImpl) FeeHistory(blockCount uint64, lastBlock rpc.BlockNumber,
-	rewardPercentiles []float64,
-) (*rpctypes.FeeHistoryResult, error) {
-	blockEnd := int64(lastBlock)
-
-	if blockEnd < 0 {
-		blockNumber, err := b.BlockNumber()
-		if err != nil {
-			return nil, err
-		}
-		blockEnd = int64(blockNumber)
-	}
-
-	blocks := int64(blockCount)
-	maxBlockCount := int64(b.cfg.AppCfg.JSONRPC.FeeHistoryCap)
-	if blocks > maxBlockCount {
-		return nil, fmt.Errorf("FeeHistory user block count %d higher than %d", blocks, maxBlockCount)
-	}
-
-	if blockEnd+1 < blocks {
-		blocks = blockEnd + 1
-	}
-
-	blockStart := blockEnd + 1 - blocks
-	oldestBlock := (*hexutil.Big)(big.NewInt(blockStart))
-
-	reward := make([][]*hexutil.Big, blocks)
-	rewardCount := len(rewardPercentiles)
-	for i := 0; i < int(blocks); i++ {
-		reward[i] = make([]*hexutil.Big, rewardCount)
-	}
-
-	thisBaseFee := make([]*hexutil.Big, blocks+1)
-	thisGasUsedRatio := make([]float64, blocks)
-	calculateRewards := rewardCount != 0
-
-	for blockID := blockStart; blockID <= blockEnd; blockID++ {
-		index := int32(blockID - blockStart) // #nosec G701
-		// tendermint block
-		tendermintblock, err := b.CosmosBlockByNumber(rpc.BlockNumber(blockID))
-		if tendermintblock == nil {
-			return nil, err
-		}
-
-		// eth block
-		ethBlock, err := b.GetBlockByNumber(rpc.BlockNumber(blockID), true)
-		if ethBlock == nil {
-			return nil, err
-		}
-
-		// tendermint block result
-		tendermintBlockResult, err := b.CosmosBlockResultByNumber(&tendermintblock.Block.Height)
-		if tendermintBlockResult == nil {
-			b.logger.Debug("block result not found", "height", tendermintblock.Block.Height, "error", err.Error())
-			return nil, err
-		}
-
-		oneFeeHistory, err := b.processBlock(tendermintblock, &ethBlock, rewardPercentiles, tendermintBlockResult)
-		if err != nil {
-			return nil, err
-		}
-
-		// copy
-		thisBaseFee[index] = (*hexutil.Big)(oneFeeHistory.BaseFee)
-		thisBaseFee[index+1] = (*hexutil.Big)(oneFeeHistory.NextBaseFee)
-		thisGasUsedRatio[index] = oneFeeHistory.GasUsedRatio
-		if calculateRewards {
-			for j := 0; j < rewardCount; j++ {
-				reward[index][j] = (*hexutil.Big)(oneFeeHistory.Reward[j])
-				if reward[index][j] == nil {
-					reward[index][j] = (*hexutil.Big)(big.NewInt(0))
-				}
-			}
-		}
-	}
-
-	feeHistory := rpctypes.FeeHistoryResult{
-		OldestBlock:  oldestBlock,
-		BaseFee:      thisBaseFee,
-		GasUsedRatio: thisGasUsedRatio,
-	}
-
-	if calculateRewards {
-		feeHistory.Reward = reward
-	}
-
-	return &feeHistory, nil
 }
 
 func (b *BackendImpl) ChainDb() ethdb.Database {
@@ -285,17 +257,6 @@ func (b *BackendImpl) RPCGasCap() uint64 {
 
 func (b *BackendImpl) RPCEVMTimeout() time.Duration {
 	return b.cfg.RPCEVMTimeout
-}
-
-func (b *BackendImpl) RPCTxFeeCap() float64 {
-	return b.cfg.RPCTxFeeCap
-}
-
-func (b *BackendImpl) UnprotectedAllowed() bool {
-	if b.cfg.AppCfg == nil {
-		return false
-	}
-	return b.cfg.AppCfg.JSONRPC.AllowUnprotectedTxs
 }
 
 // This is copied from filters.Backend
@@ -335,45 +296,6 @@ func (b *BackendImpl) BloomStatus() (uint64, uint64) {
 func (b *BackendImpl) ServiceFilter(_ context.Context, _ *bloombits.MatcherSession) {
 }
 
-// artela rpc DebugAPI
-
-// artela rpc DebugAPI
-
-func (b *BackendImpl) Listening() bool {
-	tmClient := b.clientCtx.Client.(bftclient.Client)
-	netInfo, err := tmClient.NetInfo(b.ctx)
-	if err != nil {
-		return false
-	}
-	return netInfo.Listening
-}
-
-func (b *BackendImpl) PeerCount() hexutil.Uint {
-	tmClient := b.clientCtx.Client.(bftclient.Client)
-	netInfo, err := tmClient.NetInfo(b.ctx)
-	if err != nil {
-		return 0
-	}
-	return hexutil.Uint(len(netInfo.Peers))
-}
-
-// ClientVersion returns the current client version.
-func (b *BackendImpl) ClientVersion() string {
-	return fmt.Sprintf(
-		"%s/%s/%s/%s",
-		version.Name,
-		version.Version,
-		runtime.GOOS+"-"+runtime.GOARCH,
-		runtime.Version(),
-	)
-}
-
-// func (b *BackendImpl) GetBlockContext(
-// 	_ context.Context, header *ethtypes.Header,
-// ) *vm.BlockContext {
-// 	return nil
-// }
-
 func (b *BackendImpl) BaseFee(blockRes *tmrpctypes.ResultBlockResults) (*big.Int, error) {
 	// return BaseFee if London hard fork is activated and feemarket is enabled
 	res, err := b.queryClient.BaseFee(rpctypes.ContextWithHeight(blockRes.Height), &txs.QueryBaseFeeRequest{})
@@ -400,130 +322,6 @@ func (b *BackendImpl) BaseFee(blockRes *tmrpctypes.ResultBlockResults) (*big.Int
 	}
 
 	return res.BaseFee.BigInt(), nil
-}
-
-func (b *BackendImpl) PendingTransactions() ([]*sdktypes.Tx, error) {
-	res, err := b.clientCtx.Client.UnconfirmedTxs(b.ctx, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	result := make([]*sdktypes.Tx, 0, len(res.Txs))
-	for _, txBz := range res.Txs {
-		tx, err := b.clientCtx.TxConfig.TxDecoder()(txBz)
-		if err != nil {
-			return nil, err
-		}
-		result = append(result, &tx)
-	}
-
-	return result, nil
-}
-
-func (b *BackendImpl) PendingTransactionsCount() (int, error) {
-	res, err := b.clientCtx.Client.UnconfirmedTxs(b.ctx, nil)
-	if err != nil {
-		return 0, err
-	}
-	return res.Total, nil
-}
-
-func (b *BackendImpl) GetResendArgs(args ethapi.TransactionArgs, gasPrice *hexutil.Big, gasLimit *hexutil.Uint64) (ethapi.TransactionArgs, error) {
-	chainID, err := types.ParseChainID(b.clientCtx.ChainID)
-	if err != nil {
-		return ethapi.TransactionArgs{}, err
-	}
-
-	cfg := b.ChainConfig()
-	if cfg == nil {
-		header, err := b.CurrentHeader()
-		if err != nil {
-			return ethapi.TransactionArgs{}, err
-		}
-		cfg = support.DefaultChainConfig().EthereumConfig(header.Number.Int64(), chainID)
-	}
-
-	// use the latest signer for the new tx
-	signer := ethtypes.LatestSigner(cfg)
-
-	matchTx := args.ToTransaction()
-
-	// Before replacing the old transaction, ensure the _new_ transaction fee is reasonable.
-	price := matchTx.GasPrice()
-	if gasPrice != nil {
-		price = gasPrice.ToInt()
-	}
-	gas := matchTx.Gas()
-	if gasLimit != nil {
-		gas = uint64(*gasLimit)
-	}
-	if err := rpctypes.CheckTxFee(price, gas, b.RPCTxFeeCap()); err != nil {
-		return ethapi.TransactionArgs{}, err
-	}
-
-	pending, err := b.PendingTransactions()
-	if err != nil {
-		return ethapi.TransactionArgs{}, err
-	}
-
-	for _, tx := range pending {
-		wantSigHash := signer.Hash(matchTx)
-
-		// TODO, wantSigHash?
-		msg, err := txs.UnwrapEthereumMsg(tx, wantSigHash)
-		if err != nil {
-			// not ethereum tx
-			continue
-		}
-
-		pendingTx := msg.AsTransaction()
-		pFrom, err := ethtypes.Sender(signer, pendingTx)
-		if err != nil {
-			continue
-		}
-
-		if pFrom == *args.From && signer.Hash(pendingTx) == wantSigHash {
-			// Match. Re-sign and send the transaction.
-			if gasPrice != nil && (*big.Int)(gasPrice).Sign() != 0 {
-				args.GasPrice = gasPrice
-			}
-			if gasLimit != nil && *gasLimit != 0 {
-				args.Gas = gasLimit
-			}
-
-			return args, nil
-		}
-	}
-
-	return ethapi.TransactionArgs{}, fmt.Errorf("transaction %s not found", matchTx.Hash().String())
-}
-
-func (b *BackendImpl) GasPrice(ctx context.Context) (*hexutil.Big, error) {
-	var (
-		result *big.Int
-		err    error
-	)
-	if head, err := b.CurrentHeader(); err == nil && head.BaseFee != nil {
-		result, err = b.SuggestGasTipCap(head.BaseFee)
-		if err != nil {
-			return nil, err
-		}
-		result = result.Add(result, head.BaseFee)
-	} else {
-		result = big.NewInt(b.RPCMinGasPrice())
-	}
-
-	// return at least GlobalMinGasPrice from FeeMarket module
-	minGasPrice, err := b.GlobalMinGasPrice()
-	if err != nil {
-		return nil, err
-	}
-	minGasPriceInt := minGasPrice.TruncateInt().BigInt()
-	if result.Cmp(minGasPriceInt) < 0 {
-		result = minGasPriceInt
-	}
-
-	return (*hexutil.Big)(result), nil
 }
 
 func (b *BackendImpl) RPCMinGasPrice() int64 {
@@ -564,133 +362,4 @@ func (b *BackendImpl) RPCFilterCap() int32 {
 // RPCLogsCap defines the max number of results can be returned from single `eth_getLogs` query.
 func (b *BackendImpl) RPCLogsCap() int32 {
 	return b.appConf.JSONRPC.LogsCap
-}
-
-func (b *BackendImpl) Syncing() (interface{}, error) {
-	status, err := b.clientCtx.Client.Status(b.ctx)
-	if err != nil {
-		return false, err
-	}
-
-	if !status.SyncInfo.CatchingUp {
-		return false, nil
-	}
-
-	return map[string]interface{}{
-		"startingBlock": hexutil.Uint64(status.SyncInfo.EarliestBlockHeight),
-		"currentBlock":  hexutil.Uint64(status.SyncInfo.LatestBlockHeight),
-		// "highestBlock":  nil, // NA
-		// "pulledStates":  nil, // NA
-		// "knownStates":   nil, // NA
-	}, nil
-}
-func (b *BackendImpl) GetCoinbase() (sdktypes.AccAddress, error) {
-	node, err := b.clientCtx.GetNode()
-	if err != nil {
-		return nil, err
-	}
-
-	status, err := node.Status(b.ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	req := &txs.QueryValidatorAccountRequest{
-		ConsAddress: sdktypes.ConsAddress(status.ValidatorInfo.Address).String(),
-	}
-
-	res, err := b.queryClient.ValidatorAccount(b.ctx, req)
-	if err != nil {
-		return nil, err
-	}
-
-	address, _ := sdktypes.AccAddressFromBech32(res.AccountAddress) // #nosec G703
-	return address, nil
-}
-
-// GetProof returns an account object with proof and any storage proofs
-func (b *BackendImpl) GetProof(address common.Address, storageKeys []string, blockNrOrHash rpctypes.BlockNumberOrHash) (*rpctypes.AccountResult, error) {
-	numberOrHash := rpc.BlockNumberOrHash{
-		BlockNumber:      (*rpc.BlockNumber)(blockNrOrHash.BlockNumber),
-		BlockHash:        blockNrOrHash.BlockHash,
-		RequireCanonical: false,
-	}
-	blockNum, err := b.blockNumberFromCosmos(numberOrHash)
-	if err != nil {
-		return nil, err
-	}
-
-	height := blockNum.Int64()
-
-	_, err = b.CosmosBlockByNumber(blockNum)
-	if err != nil {
-		// the error message imitates geth behavior
-		return nil, fmt.Errorf("block not valid, %v", err)
-	}
-	ctx := rpctypes.ContextWithHeight(height)
-
-	// if the height is equal to zero, meaning the query condition of the block is either "pending" or "latest"
-	if height == 0 {
-		bn, err := b.BlockNumber()
-		if err != nil {
-			return nil, err
-		}
-
-		if bn > math.MaxInt64 {
-			return nil, fmt.Errorf("not able to query block number greater than MaxInt64")
-		}
-
-		height = int64(bn) // #nosec G701 -- checked for int overflow already
-	}
-
-	clientCtx := b.clientCtx.WithHeight(height)
-
-	// query storage proofs
-	storageProofs := make([]rpctypes.StorageResult, len(storageKeys))
-
-	for i, key := range storageKeys {
-		hexKey := common.HexToHash(key)
-		valueBz, proof, err := b.queryClient.GetProof(clientCtx, evmtypes.StoreKey, evmtypes.StateKey(address, hexKey.Bytes()))
-		if err != nil {
-			return nil, err
-		}
-
-		storageProofs[i] = rpctypes.StorageResult{
-			Key:   key,
-			Value: (*hexutil.Big)(new(big.Int).SetBytes(valueBz)),
-			Proof: utils.GetHexProofs(proof),
-		}
-	}
-
-	// query EVM account
-	req := &txs.QueryAccountRequest{
-		Address: address.String(),
-	}
-
-	res, err := b.queryClient.Account(ctx, req)
-	if err != nil {
-		return nil, err
-	}
-
-	// query account proofs
-	accountKey := authtypes.AddressStoreKey(sdktypes.AccAddress(address.Bytes()))
-	_, proof, err := b.queryClient.GetProof(clientCtx, authtypes.StoreKey, accountKey)
-	if err != nil {
-		return nil, err
-	}
-
-	balance, ok := sdkmath.NewIntFromString(res.Balance)
-	if !ok {
-		return nil, errors.New("invalid balance")
-	}
-
-	return &rpctypes.AccountResult{
-		Address:      address,
-		AccountProof: utils.GetHexProofs(proof),
-		Balance:      (*hexutil.Big)(balance.BigInt()),
-		CodeHash:     common.HexToHash(res.CodeHash),
-		Nonce:        hexutil.Uint64(res.Nonce),
-		StorageHash:  common.Hash{}, // NOTE: Evmos doesn't have a storage hash. TODO: implement?
-		StorageProof: storageProofs,
-	}, nil
 }
