+++ ethereum/rpc/tx_pool.go
@@ -1,0 +1,9 @@
+package rpc
+
+func (b *BackendImpl) PendingTransactionsCount() (int, error) {
+	res, err := b.clientCtx.Client.UnconfirmedTxs(b.ctx, nil)
+	if err != nil {
+		return 0, err
+	}
+	return res.Total, nil
+}
